Working visitations counting endpoints for visitations by target month, in the last six months and for today

+++ recipes-server/src/main/java/recepiesserver/recipesserver/controllers/RecipeController.java
@@ -87,8 +87,8 @@ public class RecipeController {
     }
 
     @PostMapping("/{id}/visitations")
-    public ResponseEntity<Integer> recordVisitation(@PathVariable Long id) {
-        int incrementedVisitations = this.recipeService.incrementRecipeVisitations(id);
+    public ResponseEntity<Long> recordVisitation(@PathVariable Long id) {
+        long incrementedVisitations = this.recipeService.incrementRecipeVisitations(id);
         return ResponseEntity.ok().body(incrementedVisitations);
     }
 }
